requirements.txt
numpy==1.18.1
pandas==1.0.1
psycopg2-binary==2.8.4
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
